+++ CHANGELOG.md
@@ -27,7 +27,7 @@ Additional labels for pre-release metadata:
 * rc.x: stable release candidate.
 
 
-0.2.3 - Unreleased
+0.2.3 - 25-11-2021
 ------------------
 
 Changed
@@ -36,6 +36,7 @@ Changed
 * Bump "wasmtime" from v0.30 to v0.31.
 * Added smartcontract events to tx receipt
 * Added smartcontracts events subscription
+* Added sha256 host function
 
 0.2.2 - 05-11-2021
 ------------------